Add files via upload
This is the code that I have made for a dht11, arduino uno r4, lcd display, and with an IFTTT automation incorporated. I have made it so it is used on the IOT cloud for arduino, it is C++ language and it measures temperature and humidity with the dht sensor, it will trigger a notification to come to your phone when the temperature is over a set point that will trigger a notification.